pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "energy-trading-quant-tools"
version = "1.0.0"
description = "Institutional-grade full-stack quant platform for energy trading."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Gencer Sarp Mert"}]
dependencies = [
  # Core scientific stack
  "numpy>=1.26",
  "pandas>=2.1",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  # Async/API
  "fastapi>=0.110.0",
  "uvicorn>=0.27",
  "pydantic>=2.5",
  "websockets>=12.0",
  # Messaging / streaming
  "kafka-python>=2.0",
  # Optimization
  "pulp>=2.8",
  # Data access
  "requests>=2.31",
  # ML / forecasting
  "hmmlearn>=0.3.0",
  "lightgbm>=4.0",
  "joblib>=1.3",
  # Data processing
  "polars>=0.20",
  # Visualization / dashboard
  "streamlit>=1.30",
  "plotly>=5.18",
  # Time series
  "statsmodels>=0.14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "ruff>=0.4",
  "mypy>=1.8",
  "jupyter>=1.0",
  "ipykernel>=6.0",
]
redis = [
  "redis>=5.0",
]
ray = [
  "ray>=2.9",
]
torch = [
  "torch>=2.0",
]
full = [
  "redis>=5.0",
  "ray>=2.9",
  "statsmodels>=0.14",
  "plotly>=5.18",
]

[project.scripts]
energytrader = "energytrading.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q --import-mode=importlib"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]